cli: extract the file-rewriting engine into a rewrite module
The engine that powers filter/compress/decompress had grown inside
commands::filter, and other commands reached into it (commands::filter::
run_transcode, super::filter::summary_supports_indexed_transcode). Move it
into a dedicated crate::rewrite module so the commands are thin adapters over
one engine.

This is a mechanical relocation with no behavior change:
- rust/cli/src/rewrite/mod.rs holds the moved engine (options, run, the
  read/select/place pipeline, and its tests).
- commands/filter.rs is reduced to the FilterCommand handler calling
  rewrite::run; compress/decompress build RewriteOptions and call rewrite::run;
  sort and merge reference crate::rewrite helpers.
- The transcode vocabulary is renamed to match the module: run_transcode -> run,
  TranscodeCommandOptions -> RewriteOptions, summary_supports_indexed_transcode
  -> summary_supports_indexed_read, read_summary_for_indexed_transcode ->
  read_indexed_summary, build_filter_options_from_transcode_options ->
  resolve_options.

No sort port, no --order, no memory/reindex work; those follow in later PRs.
