+++ rust/cli/src/commands/compress.rs
@@ -1,16 +1,16 @@
 use anyhow::Result;
 
 use crate::cli::CompressCommand;
-use crate::commands::filter::{self, TranscodeCommandOptions};
 use crate::context::CommandContext;
+use crate::rewrite::{self, RewriteOptions};
 
 pub fn run(ctx: &CommandContext, args: CompressCommand) -> Result<()> {
-    let options = TranscodeCommandOptions::new(args.file, args.output, args.chunk_size)
+    let options = RewriteOptions::new(args.file, args.output, args.chunk_size)
         .compression(args.compression)
         .use_chunks(true)
         .include_metadata(true)
         .include_attachments(true);
-    filter::run_transcode(
+    rewrite::run(
         options,
         crate::source::SourceOptions::new(ctx.allow_remote_scan()),
     )
